+++ demisto_sdk/commands/common/update_id_set.py
@@ -28,8 +28,9 @@ from demisto_sdk.commands.common.constants import (CLASSIFIERS_DIR,
                                                    INCIDENT_TYPES_DIR,
                                                    INDICATOR_FIELDS_DIR,
                                                    INDICATOR_TYPES_DIR,
-                                                   LAYOUTS_DIR, MAPPERS_DIR,
-                                                   REPORTS_DIR, SCRIPTS_DIR,
+                                                   LAYOUTS_DIR, LISTS_DIR,
+                                                   MAPPERS_DIR, REPORTS_DIR,
+                                                   SCRIPTS_DIR,
                                                    TEST_PLAYBOOKS_DIR,
                                                    WIDGETS_DIR, FileType)
 from demisto_sdk.commands.common.tools import (LOG_COLORS, find_type, get_json,
@@ -41,12 +42,12 @@ from demisto_sdk.commands.unify.yml_unifier import YmlUnifier
 CONTENT_ENTITIES = ['Integrations', 'Scripts', 'Playbooks', 'TestPlaybooks', 'Classifiers',
                     'Dashboards', 'IncidentFields', 'IncidentTypes', 'IndicatorFields', 'IndicatorTypes',
                     'Layouts', 'Reports', 'Widgets', 'Mappers', 'Packs', 'GenericTypes',
-                    'GenericFields', 'GenericModules', 'GenericDefinitions']
+                    'GenericFields', 'GenericModules', 'GenericDefinitions', 'Lists']
 
 ID_SET_ENTITIES = ['integrations', 'scripts', 'playbooks', 'TestPlaybooks', 'Classifiers',
                    'Dashboards', 'IncidentFields', 'IncidentTypes', 'IndicatorFields', 'IndicatorTypes',
                    'Layouts', 'Reports', 'Widgets', 'Mappers', 'GenericTypes', 'GenericFields', 'GenericModules',
-                   'GenericDefinitions']
+                   'GenericDefinitions', 'Lists']
 
 BUILT_IN_FIELDS = [
     "name",
@@ -142,6 +143,30 @@ def build_tasks_graph(playbook_data):
         found_new_tasks = graph.number_of_nodes() > current_number_of_nodes
 
     return graph
+
+
+def get_lists_names_from_playbook(data_dictionary: dict, graph: networkx.DiGraph) -> tuple:
+    lists_names = set()
+    lists_names_skippable = set()
+    tasks = data_dictionary.get('tasks', {})
+    lists_tasks_scripts = ['Builtin|||setList', 'Builtin|||getList']
+    for task_id, task in tasks.items():
+        script = task.get('task', {}).get('script')
+        if script in lists_tasks_scripts:
+            list_name = task.get('scriptarguments', {}).get('listName', {}).get('simple')
+
+            try:
+                skippable = not graph.nodes[task_id]['mandatory']
+            except KeyError:
+                # if task id not in the graph - the task is unreachable.
+                print_error(f'{data_dictionary["id"]}: task {task_id} is not connected')
+                continue
+            if list_name:
+                lists_names.add(list_name)
+                if skippable:
+                    lists_names_skippable.add(list_name)
+
+    return list(lists_names), list(lists_names_skippable)
 
 
 def get_task_ids_from_playbook(param_to_enrich_by: str, data_dict: dict, graph: networkx.DiGraph) -> tuple:
@@ -452,9 +477,10 @@ def get_playbook_data(file_path: str) -> dict:
                                                                                           data_dictionary,
                                                                                           graph
                                                                                           )
+    implementing_lists, implementing_lists_skippable = get_lists_names_from_playbook(data_dictionary, graph)
     command_to_integration, command_to_integration_skippable = get_commands_from_playbook(data_dictionary)
     skippable_tasks = (implementing_scripts_skippable + implementing_playbooks_skippable +
-                       command_to_integration_skippable)
+                       command_to_integration_skippable + implementing_lists_skippable)
     pack = get_pack_name(file_path)
     dependent_incident_fields, dependent_indicator_fields = get_dependent_incident_and_indicator_fields(data_dictionary)
 
@@ -483,6 +509,8 @@ def get_playbook_data(file_path: str) -> dict:
         playbook_data['filters'] = filters
     if transformers:
         playbook_data['transformers'] = transformers
+    if implementing_lists:
+        playbook_data['lists'] = implementing_lists
     return {id_: playbook_data}
 
 
@@ -1108,7 +1136,7 @@ def process_generic_items(file_path: str, print_logs: bool,
     """
     Process a generic field JSON file
     Args:
-        file_path: The file path from pbject field folder
+        file_path: The file path from object field folder
         print_logs: Whether to print logs to stdout.
         generic_types_list: List of all the generic types in the system.
         generic_modules_list: List of all the generic modules in the system.
@@ -1135,7 +1163,7 @@ def process_generic_items(file_path: str, print_logs: bool,
 def process_general_items(file_path: str, print_logs: bool, expected_file_types: Tuple[FileType],
                           data_extraction_func: Callable) -> list:
     """
-    Process a generic item file.
+    Process a general item file.
     expected file in one of the following:
     * classifier
     * incident type
@@ -1146,6 +1174,7 @@ def process_general_items(file_path: str, print_logs: bool, expected_file_types:
     * playbook
     * report
     * widget
+    * list
 
     Args:
         file_path: The file path from an item folder
@@ -1378,6 +1407,17 @@ def get_generic_module_data(path):
     return {id_: data}
 
 
+def get_list_data(path: str):
+    json_data = get_json(path)
+    data = create_common_entity_data(path=path,
+                                     name=json_data.get('name'),
+                                     to_version=json_data.get('toVersion'),
+                                     from_version=json_data.get('fromVersion'),
+                                     pack=get_pack_name(path))
+
+    return {json_data.get('id'): data}
+
+
 class IDSetType(Enum):
     PLAYBOOK = 'playbooks'
     INTEGRATION = 'integrations'
@@ -1551,6 +1591,7 @@ def re_create_id_set(id_set_path: Optional[str] = DEFAULT_ID_SET_PATH, pack_to_c
     generic_fields_list = []
     generic_modules_list = []
     generic_definitions_list = []
+    lists_list = []
     packs_dict: Dict[str, Dict] = {}
 
     pool = Pool(processes=int(cpu_count()))
@@ -1741,6 +1782,18 @@ def re_create_id_set(id_set_path: Optional[str] = DEFAULT_ID_SET_PATH, pack_to_c
 
         progress_bar.update(1)
 
+        if 'Lists' in objects_to_create:
+            print_color("\nStarting iteration over Lists", LOG_COLORS.GREEN)
+            for arr in pool.map(partial(process_general_items,
+                                        print_logs=print_logs,
+                                        expected_file_types=(FileType.LISTS,),
+                                        data_extraction_func=get_list_data,
+                                        ),
+                                get_general_paths(LISTS_DIR, pack_to_create)):
+                lists_list.extend(arr)
+
+        progress_bar.update(1)
+
         if 'GenericDefinitions' in objects_to_create:
             print_color("\nStarting iteration over Generic Definitions", LOG_COLORS.GREEN)
             print_color(f"pack to create: {pack_to_create}", LOG_COLORS.YELLOW)
@@ -1811,6 +1864,7 @@ def re_create_id_set(id_set_path: Optional[str] = DEFAULT_ID_SET_PATH, pack_to_c
     new_ids_dict['GenericFields'] = sort(generic_fields_list)
     new_ids_dict['GenericModules'] = sort(generic_modules_list)
     new_ids_dict['GenericDefinitions'] = sort(generic_definitions_list)
+    new_ids_dict['Lists'] = sort(lists_list)
 
     exec_time = time.time() - start_time
     print_color("Finished the creation of the id_set. Total time: {} seconds".format(exec_time), LOG_COLORS.GREEN)
@@ -1881,13 +1935,8 @@ def has_duplicate(id_set_subset_list, id_to_check, object_type=None, print_logs=
         if object_type == 'Layouts':
             if dict1.get('kind', '') != dict2.get('kind', ''):
                 return False
-            if dict1.get('typeId', '') != dict2.get('typeId', ''):
+            if dict1.get('typeID', '') != dict2.get('typeID', ''):
                 return False
-
-        # If they have the same pack name they actually the same entity.
-        # Added to support merge between two id-sets that contain the same pack.
-        if dict1.get('pack') == dict2.get('pack'):
-            return False
 
         # A: 3.0.0 - 3.6.0
         # B: 3.5.0 - 4.5.0
@@ -1899,9 +1948,9 @@ def has_duplicate(id_set_subset_list, id_to_check, object_type=None, print_logs=
             dict2_from_version <= dict1_from_version < dict2_to_version,  # will catch (C, B), (B, A), (C, A)
             dict2_from_version < dict1_to_version <= dict2_to_version,  # will catch (C, B), (C, A)
         ]):
-            print_warning('The following {} have the same ID ({}) and their versions overlap: '
-                          '"1.{}-{}", "2.{}-{}".'.format(object_type, id_to_check, dict1_from_version, dict1_to_version,
-                                                         dict2_from_version, dict2_to_version))
+            print_warning(f'The following {object_type} have the same ID ({id_to_check}) and their versions overlap: '
+                          f'1) "{dict1_from_version}-{dict1_to_version}", '
+                          f'2) "{dict2_from_version}-{dict2_to_version}".')
             return True
 
         if print_logs and dict1.get('name') != dict2.get('name'):