+++ demisto_sdk/commands/generate_outputs/json_to_outputs/tests/json_to_outputs_test.py
@@ -5,11 +5,11 @@ from typing import Optional
 import pytest
 
 from demisto_sdk.commands.common.tools import run_command
-from demisto_sdk.commands.json_to_outputs.json_to_outputs import (
+from demisto_sdk.commands.generate_outputs.json_to_outputs.json_to_outputs import (
     determine_type, json_to_outputs, parse_json)
 
 DUMMY_FIELD_DESCRIPTION = "dummy field description"
-TEST_PATH = Path('demisto_sdk/commands/json_to_outputs/tests')
+TEST_PATH = Path('demisto_sdk/commands/generate_outputs/json_to_outputs/tests')
 
 
 def git_path() -> str:
@@ -243,7 +243,7 @@ def test_json_to_outputs__description_dictionary(tmpdir, description_argument: O
     temp_json_input_path.write(json.dumps(dummy_integration_output))
 
     json_to_outputs(command='jsonToOutputs',
-                    input=str(temp_json_input_path),
+                    json=str(temp_json_input_path),
                     prefix='Test',
                     output=output,
                     descriptions=description_argument)
@@ -282,7 +282,7 @@ def test_json_to_outputs__description_file(tmpdir):
     temp_description_file.write(json.dumps(dummy_description_dictionary))
 
     json_to_outputs(command='jsonToOutputs',
-                    input=str(temp_json_input_path),
+                    json=str(temp_json_input_path),
                     prefix='Test',
                     output=output,
                     descriptions=temp_description_file)